fix(api): use double instead of float for jsonb polygon points
@JdbcTypeCode(SqlTypes.JSON) deserializes JSON numbers as double at runtime.
List<List<Float>> caused a ClassCastException in Jackson's FloatSerializer
because JVM type erasure allowed double values to enter the list undetected.

+++ src/main/kotlin/com/hanmaum/dn/app/features/floorplan/api/v1/dto/FloorPlanDtos.kt
@@ -11,5 +11,5 @@ data class RoomResponse(
     val floorId: String,
     val name: String,
     val description: String,
-    val points: List<List<Float>>,
+    val points: List<List<Double>>,
 )
